+++ terra-rust-api/src/client/tx.rs
@@ -1,8 +1,11 @@
+use reqwest::StatusCode;
 //use crate::client::core_types::Msg;
 use crate::client::tx_types::{
     TXResultAsync, TXResultBlock, TXResultSync, TxEstimate, TxFeeResult,
 };
 use crate::core_types::{Coin, StdSignMsg, StdSignature, StdTx};
+use crate::errors::TerraRustAPIError;
+use crate::errors::TerraRustAPIError::TXNotFound;
 use crate::messages::Message;
 use crate::{LCDResult, Terra};
 
@@ -39,7 +42,10 @@ impl<'a> TX<'a> {
     ) -> anyhow::Result<TXResultSync> {
         let std_tx: StdTx = StdTx::from_StdSignMsg(std_sign_msg, sigs, "sync");
         //    let js_sig = serde_json::to_string(&std_tx)?;
-        log::info!("{}", serde_json::to_string(&std_tx)?);
+        log::info!("TX broadcast #messages ={}", &std_tx.tx.msg.len());
+        if self.terra.debug {
+            log::debug!("{}", serde_json::to_string(&std_tx)?);
+        }
         let response = self
             .terra
             .post_cmd::<StdTx, TXResultSync>("/txs", &std_tx)
@@ -63,12 +69,49 @@ impl<'a> TX<'a> {
         Ok(response)
     }
     /// get TX result
-    pub async fn get(&self, hash: &str) -> anyhow::Result<TXResultBlock> {
+    pub async fn get(&self, hash: &str) -> Result<TXResultBlock, TerraRustAPIError> {
         let resp = self
             .terra
             .send_cmd::<TXResultBlock>(&format!("/txs/{}", hash), None)
             .await?;
         Ok(resp)
+    }
+    /// get TX result
+    pub async fn get_and_wait(
+        &self,
+        hash: &str,
+        max_times: usize,
+        sleep_amount: tokio::time::Duration,
+    ) -> Result<TXResultBlock, TerraRustAPIError> {
+        let mut times = 0;
+        while times < max_times {
+            let tx = self.get(hash).await;
+            match tx {
+                Ok(tx_response) => return Ok(tx_response),
+                Err(e) => {
+                    times += 1;
+                    match &e {
+                        TerraRustAPIError::TerraLCDResponse(statuscode, out) => {
+                            if statuscode == &StatusCode::NOT_FOUND {
+                                log::info!(
+                                    "Transaction not applied .. retry #{} sleeping 1s",
+                                    times
+                                );
+                                tokio::time::sleep(sleep_amount).await;
+                            } else {
+                                log::error!("Invalid Response TX: {} {}", statuscode, out);
+                                break;
+                            }
+                        }
+                        _ => {
+                            log::error!("Invalid Response TX: {}", e);
+                            break;
+                        }
+                    }
+                }
+            }
+        }
+        Err(TXNotFound(max_times))
     }
     /// Estimate the StdFee structure based on the gas used
     pub async fn estimate_fee(
@@ -86,7 +129,15 @@ impl<'a> TX<'a> {
             gas_prices,
         );
 
-        log::info!("Estimate Transaction = {}", serde_json::to_string(&tx_est)?);
+        if self.terra.debug {
+            log::info!("Estimate Transaction = {}", serde_json::to_string(&tx_est)?);
+        } else {
+            log::debug!(
+                "Estimate Transaction = {:#?} #messages={}",
+                tx_est.base_req,
+                tx_est.msgs.len()
+            )
+        }
         let resp = self
             .terra
             .post_cmd::<TxEstimate, LCDResult<TxFeeResult>>("/txs/estimate_fee", &tx_est)